XP-3998 Replace three details panels with one - Separated slidable panels implementation from details view that contains widget - Refactored details panel to separate sliding panels functionality from widgets view functionality - Renamed DetailsPanel into DetailsView- now details view is responsible only for rendering widgets, not hiding/showing itself
