+++ modules/admin/admin-ui/src/main/resources/web/admin/apps/content-studio/js/app/view/detail/DockedDetailsPanel.ts
@@ -1,0 +1,33 @@
+import "../../../api.ts";
+import {DetailsPanel, DETAILS_PANEL_TYPE} from "./DetailsPanel";
+import {DetailsView} from "./DetailsView";
+
+import ResponsiveManager = api.ui.responsive.ResponsiveManager;
+
+export class DockedDetailsPanel extends DetailsPanel {
+
+    constructor(detailsView: DetailsView) {
+        super(detailsView);
+        this.setDoOffset(false);
+        this.addClass("docked-details-panel");
+    }
+
+    protected subscribeOnEvents() {
+        this.onPanelSizeChanged(() => this.detailsView.setDetailsContainerHeight());
+
+        this.onShown(() => {
+            if (!!this.getItem()) {
+                setTimeout(() => this.detailsView.updateActiveWidget(), 250); // small delay so that isVisibleOrAboutToBeVisible() check detects width change
+            }
+        });
+    }
+
+    public isVisibleOrAboutToBeVisible(): boolean {
+        return this.getHTMLElement().clientWidth > 0;
+    }
+
+    public getType(): DETAILS_PANEL_TYPE {
+        return DETAILS_PANEL_TYPE.DOCKED;
+    }
+}
+
