+++ modules/admin/admin-ui/src/main/resources/web/admin/apps/content-studio/js/app/view/detail/WidgetView.ts
@@ -4,8 +4,9 @@ import ViewItem = api.app.view.ViewItem;
 import ContentSummaryAndCompareStatus = api.content.ContentSummaryAndCompareStatus;
 import RenderingMode = api.rendering.RenderingMode;
 import Widget = api.content.Widget;
-import {DetailsPanel} from "./DetailsPanel";
+import {DetailsView} from "./DetailsView";
 import {WidgetItemView} from "./WidgetItemView";
+import {ActiveDetailsPanelManager} from "../../view/detail/ActiveDetailsPanelManager";
 
 export class WidgetView extends api.dom.DivEl {
 
@@ -13,7 +14,7 @@ export class WidgetView extends api.dom.DivEl {
 
     private widgetItemViews: WidgetItemView[];
 
-    private detailsPanel: DetailsPanel;
+    private detailsView: DetailsView;
 
     private widget: Widget;
 
@@ -30,7 +31,7 @@ export class WidgetView extends api.dom.DivEl {
     constructor(builder: WidgetViewBuilder) {
         super("widget-view " + (builder.widget ? "external-widget" : "internal-widget"));
 
-        this.detailsPanel = builder.detailsPanel;
+        this.detailsView = builder.detailsView;
         this.widgetName = builder.name;
         this.widgetItemViews = builder.widgetItemViews;
         this.widget = builder.widget;
@@ -62,12 +63,12 @@ export class WidgetView extends api.dom.DivEl {
 
     private handleRerenderOnResize() {
         var updateWidgetItemViewsHandler = () => {
-            var containerWidth = this.detailsPanel.getEl().getWidth();
-            if (this.detailsPanel.getItem() && containerWidth !== this.containerWidth) {
+            var containerWidth = this.detailsView.getEl().getWidth();
+            if (this.detailsView.getItem() && containerWidth !== this.containerWidth) {
                 this.updateWidgetItemViews(true);
             }
         }
-        this.detailsPanel.onPanelSizeChanged(() => {
+        this.detailsView.onPanelSizeChanged(() => {
             if (this.isActive()) {
                 updateWidgetItemViewsHandler();
             } else {
@@ -85,7 +86,7 @@ export class WidgetView extends api.dom.DivEl {
     }
 
     private getFullUrl(url: string) {
-        return url + "/" + this.detailsPanel.getEl().getWidth();
+        return url + "/" + this.detailsView.getEl().getWidth();
     }
 
     private updateCustomWidgetItemViews(force: boolean = false): wemQ.Promise<any>[] {
@@ -100,11 +101,11 @@ export class WidgetView extends api.dom.DivEl {
     }
 
     public updateWidgetItemViews(force: boolean = false): wemQ.Promise<any> {
-        var content = this.detailsPanel.getItem(),
+        var content = this.detailsView.getItem(),
             promises = [];
 
         if (this.widgetShouldBeUpdated(force)) {
-            this.detailsPanel.showLoadMask();
+            this.detailsView.showLoadMask();
             this.content = content;
 
             if (this.isUrlBased()) {
@@ -116,19 +117,19 @@ export class WidgetView extends api.dom.DivEl {
             }
         }
 
-        this.containerWidth = this.detailsPanel.getEl().getWidth();
-        return wemQ.all(promises).finally(() => this.detailsPanel.hideLoadMask());
+        this.containerWidth = this.detailsView.getEl().getWidth();
+        return wemQ.all(promises).finally(() => this.detailsView.hideLoadMask());
     }
 
     private widgetShouldBeUpdated(force: boolean = false): boolean {
-        var content = this.detailsPanel.getItem();
-        return content && this.detailsPanel.isVisibleOrAboutToBeVisible() &&
+        var content = this.detailsView.getItem();
+        return content && ActiveDetailsPanelManager.getActiveDetailsPanel().isVisibleOrAboutToBeVisible() &&
                (force || !api.ObjectHelper.equals(this.content, content) || (this.isUrlBased() && this.url !== this.getWidgetUrl()));
     }
 
     private createDefaultWidgetItemView() {
         this.widgetItemViews.push(new WidgetItemView());
-        if (this.detailsPanel.getItem()) {
+        if (this.detailsView.getItem()) {
             this.updateWidgetItemViews();
         }
     }
@@ -145,24 +146,6 @@ export class WidgetView extends api.dom.DivEl {
         });
 
         return wemQ.all(layoutTasks);
-    }
-
-    private calcHeight(): number {
-        var originalHeight = this.getEl().getHeight();
-        if (originalHeight == 0) {
-            // prevent jitter if widget is collapsed
-            this.setVisible(false);
-        }
-        this.getEl().setHeight("auto");
-        var height = this.getEl().getHeight();
-        this.getEl().setHeightPx(originalHeight);
-        if (originalHeight == 0) {
-            this.setVisible(true);
-        }
-        if (WidgetView.debug) {
-            console.debug('WidgetView.calcHeight: ', height, 'originalHeight: ', originalHeight);
-        }
-        return height;
     }
 
     getWidgetName(): string {
@@ -194,7 +177,7 @@ export class WidgetView extends api.dom.DivEl {
         if (this.isActive()) {
             return;
         }
-        this.detailsPanel.setActiveWidget(this);
+        this.detailsView.setActiveWidget(this);
         this.notifyActivated();
         this.slideIn();
     }
@@ -203,12 +186,12 @@ export class WidgetView extends api.dom.DivEl {
         if (WidgetView.debug) {
             console.debug('WidgetView.setInactive: ', this.getWidgetName());
         }
-        this.detailsPanel.resetActiveWidget();
+        this.detailsView.resetActiveWidget();
         this.slideOut();
     }
 
     private isActive() {
-        return this.detailsPanel.getActiveWidget() == this;
+        return this.detailsView.getActiveWidget() == this;
     }
 
     private hasDynamicHeight(): boolean {
@@ -229,10 +212,6 @@ export class WidgetView extends api.dom.DivEl {
 
     private isUrlBased(): boolean {
         return !!this.widget && !!this.widget.getUrl();
-    }
-
-    public getDetailsPanel(): DetailsPanel {
-        return this.detailsPanel;
     }
 
     notifyActivated() {
@@ -258,7 +237,7 @@ export class WidgetViewBuilder {
 
     name: string;
 
-    detailsPanel: DetailsPanel;
+    detailsView: DetailsView;
 
     widgetItemViews: WidgetItemView[] = [];
 
@@ -269,8 +248,8 @@ export class WidgetViewBuilder {
         return this;
     }
 
-    public setDetailsPanel(detailsPanel: DetailsPanel): WidgetViewBuilder {
-        this.detailsPanel = detailsPanel;
+    public setDetailsView(detailsView: DetailsView): WidgetViewBuilder {
+        this.detailsView = detailsView;
         return this;
     }
