+++ modules/admin/admin-ui/src/main/resources/web/admin/common/styles/apps/content/view/mobile-content-item-statistics-panel.less
@@ -86,6 +86,10 @@
     right: 0px;
     opacity: 0.95;
 
+    .content-summary-viewer {
+      display: none !important;
+    }
+
     > .details-container {
       left: 0px;
     }
